edit and delete InputAccount component
file InputAccountComponent in Component is edited and file SelecInputSender in Transfer is deleted
